Implement damaging combatant
Also remove some parts from json parsing that broke with the version update

+++ app/src/main/java/de/lehrbaum/initiativetracker/view/combat/client/ClientCombatantViewModel.kt
@@ -3,7 +3,7 @@ package de.lehrbaum.initiativetracker.view.combat.client
 data class ClientCombatantViewModel(
 	val id: Long,
 	val name: String,
-	val initiative: Short,
+	val initiative: Int,
 	val editMode: Boolean = false,
 	var active: Boolean = false
 ) : Comparable<ClientCombatantViewModel> {
